# VideoBackend/OGL: Prefer KHR_shader_subgroup over NV_shader_thread.