--- a/Source/Core/VideoBackends/Metal/MTLUtil.mm
+++ b/Source/Core/VideoBackends/Metal/MTLUtil.mm
@@ -386,7 +386,6 @@ static const std::string_view SUBGROUP_HELPER_HEADER = R"(
 #extension GL_KHR_shader_subgroup_ballot : enable
 
 #define SUPPORTS_SUBGROUP_REDUCTION 1
-#define CAN_USE_SUBGROUP_REDUCTION true
 #define IS_HELPER_INVOCATION gl_HelperInvocation
 #define IS_FIRST_ACTIVE_INVOCATION (subgroupElect())
 #define SUBGROUP_MIN(value) value = subgroupMin(value)