While the NV extension is totally fine, the KHR extension should be able to support more hardware.

For NVIDIA, the hardware either supports both or neither, it just needs a driver from the last two years.
For AMD, the drivers from late 2022-12 seems to bring support for the KHR extension.
For Intel, the KHR is also supported for some years.

This needs testing of BBOX on ALL GPU vendors, both behavior and performance.
And we need reporting if the two optional features are really supported on AMD.

Testing results:
- NVIDIA
  - Fermi (391.35): NOT AFFECTED. No support for GL_NV_shader_thread_shuffle nor GL_KHR_shader_subgroup
  - Ampere (528.24): MATCH. Supports both, same performance
- AMD
  - MISSING
- INTEL
  - Haswell (20.19.15.4624): NOT AFFECTED. No support for GL_NV_shader_thread_shuffle nor GL_KHR_shader_subgroup
  - Tiger Lake (31.0.101.4090): WIN. Supports only GL_KHR_shader_subgroup. All needed sub-features are supported. Small speedup as expected