@mbc07 2023-02-01
Second batch of tests, this time on a Pentium G3258 at 4.2 GHz, Windows 10, with a NVIDIA GT 440 (Fermi, driver 391.35) and a Intel HD Graphics (Gen7.5, Haswell GT1, driver 15.40.34.4624), same scene and settings as before:

* **On master (5.0-18485)**
**GT 440:** 118 FPS, up to 2x IR, performance starts decreasing at 3x IR, which is also the maximum it can go while still maintaining 60 FPS.
**HD Graphics:** 80 FPS at 1x IR, performance starts decreasing at 2x IR, which is also the maximum it can go while still maintaining 60 FPS.

* **This PR**
**GT 440:** 114 FPS, up to 2x IR, performance starts decreasing at 3x IR, which is also the maximum it can go while still maintaining 60 FPS.
**HD Graphics:** 80 FPS at 1x IR, performance starts decreasing at 2x IR, which is also the maximum it can go while still maintaining 60 FPS.

I noticed that some bounding box effects has clipping issues with the Intel HD Graphics, but this also happens on master, not exclusive to this PR:
<img src="https://user-images.githubusercontent.com/8005966/215925765-4514a1b3-bfdd-4a44-b596-f255a272dd8f.png" width=200px />
<img src="https://user-images.githubusercontent.com/8005966/215925815-595b700a-59fe-4554-9d6f-4cc502ffd23b.png" width=200px />
<img src="https://user-images.githubusercontent.com/8005966/215925864-590af5a3-8659-4f45-96fb-1117a43e1dbb.png" width=200px />