@mbc07 2023-02-01
This might be useful, it contains XML reports from GLview, listing supported extensions and other related info of the GPUs I tested:

[GLview Reports.zip](https://github.com/dolphin-emu/dolphin/files/10552867/GLview.Reports.zip)