@degasus 2023-02-01
@mbc07 Yes, this was very helpful. The XML files showed that the newer Intel driver supports everything we need and that this feature was enabled and so tested. It also shows that Fermi wasn't supported before and won't be afterwards.

However, I'm a bit doubtful about the speedup on your NVIDIA RTX 3060, it should be neglegtable as we "just" switch the GLSL extension to generate the same intrinsic.