@mbc07 2023-02-01
> However, I'm a bit doubtful about the speedup on your NVIDIA RTX 3060, it should be neglegtable as we "just" switch the GLSL extension to generate the same intrinsic.

I'll try retesting it later, this time with Optimus disabled. In my initial test I forgot to disable Optimus, so whatever was rendered on it passed through the iGPU before reaching the screen. The iGPU hasn't been particularly stable since the driver merge Intel did early January (Gen12 iGPUs and their Arc dGPUs now shares the same driver)...