@mbc07 2023-02-01
@degasus you were right. Retested the RTX 3060 with NVIDIA Optimus disabled and got 178 FPS with this PR and 176 FPS on master (min/max IR results are the same as before)...